+++ main.js
@@ -11,6 +11,9 @@ const platform = require('os').platform()
 const zmq = require('zeromq')
 const sock = zmq.socket('sub')
 
+const Manager = require('./modules/ClientBinaries/clientBinaries');
+const mgr = new Manager();
+mgr.init(true);
 
 sock.connect('tcp://127.0.0.1:30000')
 sock.subscribe('hashtx')
